Create VR - Breaking Bad.apk
This is a project to recreate an iconic environment respecting the low poly of my favorite series: Breaking Bad. Objects were modeled in Blender, developed in Unity and integrated with Google Cardboard. It was developed as a final project for the subject "Computer graphics and virtual reality" in my Computer Science college at UNA. Jun 2025.